+++ src/test/kotlin/calculator/StringAddCalculatorTest.kt
@@ -1,9 +1,8 @@
 package calculator
 
 import org.assertj.core.api.Assertions.assertThat
-import org.assertj.core.api.Assertions.assertThatExceptionOfType
+import org.junit.jupiter.api.Assertions.assertThrows
 import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.NullAndEmptySource
@@ -18,47 +17,40 @@ class StringAddCalculatorTest {
         calculator = StringAddCalculator()
     }
 
-    @DisplayName(value = "빈 문자열 또는 null 값을 입력할 경우 0을 반환해야 한다.")
     @ParameterizedTest
     @NullAndEmptySource
-    fun emptyOrNull(text: String?) {
-        assertThat(calculator.add(text)).isZero();
+    fun `빈 문자열 또는 null 값을 입력할 경우 0을 반환해야 한다`(text: String?) {
+        assertThat(calculator.add(text)).isZero
     }
 
-    @DisplayName(value = "숫자 하나를 문자열로 입력할 경우 해당 숫자를 반환한다.")
     @ParameterizedTest
     @ValueSource(strings = ["1"])
-    fun oneNumber(text: String) {
+    fun `숫자 하나를 문자열로 입력할 경우 해당 숫자를 반환한다`(text: String) {
         assertThat(calculator.add(text)).isSameAs(Integer.parseInt(text));
     }
 
-    @DisplayName(value = "숫자 두개를 쉼표(,) 구분자로 입력할 경우 두 숫자의 합을 반환한다.")
     @ParameterizedTest
     @ValueSource(strings = ["1,2"])
-    fun twoNumbers(text: String) {
+    fun `숫자 두개를 쉼표(,) 구분자로 입력할 경우 두 숫자의 합을 반환한다`(text: String) {
         assertThat(calculator.add(text)).isSameAs(3);
     }
 
-    @DisplayName(value = "구분자를 쉼표(,) 이외에 콜론(:)을 사용할 수 있다.")
     @ParameterizedTest
     @ValueSource(strings = ["1,2:3"])
-    fun colons(text: String) {
+    fun `구분자를 쉼표(,) 이외에 콜론을 사용할 수 있다`(text: String) {
         assertThat(calculator.add(text)).isSameAs(6);
     }
 
-    @DisplayName(value = "//와 \\n 문자 사이에 커스텀 구분자를 지정할 수 있다.")
     @ParameterizedTest
     @ValueSource(strings = ["//;\n1;2;3"])
-    fun customDelimiter(text: String) {
+    fun `커스텀 구분자를 지정할 수 있다`(text: String) {
         assertThat(calculator.add(text)).isSameAs(6);
     }
 
-    @DisplayName(value = "문자열 계산기에 음수를 전달하는 경우 RuntimeException 예외 처리를 한다.")
     @Test
-    fun negative() {
-        assertThatExceptionOfType(RuntimeException::class.java)
-            .isThrownBy {
-                calculator.add("-1")
-            }
+    fun `문자열 계산기에 음수를 전달하는 경우 RuntimeException 예외 처리를 한다`() {
+        assertThrows(RuntimeException::class.java) {
+            calculator.add("-1")
+        }
     }
 }
